requirements.txt
pygame==1.9.1release
pillow
numpy
